test: add tests for auth screens and composite auth form
